Created Mixin file so the code is easier to read in HaystackClient.py It allows to segment the work and separate topics Renamed readAll in read_all Rename allHistories in histories Corrected find_points
