+++ pyhaystack/client/NiagaraAXClient.py
@@ -14,7 +14,7 @@ class Connect(hc.Connect):
     This class connects to NiagaraAX and fetch haystack servlet
     A session is open and authentication will persist
     """
-    def __init__(self,url,username,password,**kwargs):
+    def __init__(self,url,username,password,*, refreshHisList = True):
         """
         Define Niagara AX specific local variables : url
         Calls the authenticate function
@@ -23,14 +23,14 @@ class Connect(hc.Connect):
         self.loginURL = self.baseURL + "login/"
         self.queryURL = self.baseURL + "haystack/"
         self.requestAbout = "about"
-        self.authenticate()
+        self.authenticate(refreshHisList = refreshHisList)
 
     def _get_kwargs(self, **kwargs):
         if 'auth' not in kwargs:
             kwargs['auth'] = (self.USERNAME, self.PASSWORD)
         return super(Connect, self)._get_kwargs(**kwargs)
 
-    def authenticate(self):
+    def authenticate(self, refreshHisList = True):
         """
         Login to the server
         Get the cookie from the server, configure headers, make a POST request with credential informations.
@@ -88,4 +88,5 @@ class Connect(hc.Connect):
             self.timezone = 'America/' + self._get_grid('read?filter=site')[0]['tz']
             print('Time Zone used : %s' % self.timezone)
             print('Connection succeed with haystack on %s (%s) running haystack version %s' %(self.serverName,self.axVersion,self.haystackVersion))
+        if refreshHisList:            
             self.refreshHisList()
