+++ pyhaystack/info.py
@@ -12,5 +12,5 @@ Project Haystack is an open source initiative to streamline working with data fr
 
 __author__ = 'Christian Tremblay, @sjlongland, @sudo-Whateverman, Igor'
 __author_email__ = 'christian.tremblay@servisys.com'
-__version__ = '0.71.0'
+__version__ = '0.71.0_issue6'
 __license__ = 'LGPL'
